+++ src/XKTModel/writeXKTModelToArrayBuffer.js
@@ -88,7 +88,7 @@ function getModelData(xktModel) {
         eachEntityPrimitiveInstancesPortion: new Uint32Array(numEntities), // For each entity, the index of the the first element of primitiveInstances used by the entity
         eachEntityMatricesPortion: new Uint32Array(numEntities), // For each entity that shares primitives, an index to its first element in data.matrices, to indicate the modeling matrix that transforms the shared primitives' Local-space vertex positions. Thios is ignored for entities that don't share primitives, because the vertex positions of non-shared primitives are pre-transformed into World-space.
 
-        eachTileAABB: new Float32Array(numTiles * 6), // For each tile, an axis-aligned bounding box
+        eachTileAABB: new Float64Array(numTiles * 6), // For each tile, an axis-aligned bounding box
         eachTileDecodeMatrix: new Float32Array(numTiles * 16), // For each tile, a position de-quantization matrix
         eachTileEntitiesPortion: new Uint32Array(numTiles) // For each tile, the index of the the first element of eachEntityId, eachEntityPrimitiveInstancesPortion and eachEntityMatricesPortion used by the tile
     };
